Minor compiler warning

+++ gtkmovelistctrl.c
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: gtkmovelistctrl.c,v 1.7 2006/04/15 01:35:20 c_anthon Exp $
+ * $Id: gtkmovelistctrl.c,v 1.8 2006/09/30 09:24:33 Superfly_Jon Exp $
  */
 
 #include "config.h"
@@ -342,13 +342,13 @@ custom_cell_renderer_movelist_get_size (GtkCellRenderer *cell,
   {
     if (x_offset)
     {
-      *x_offset = cell->xalign * (cell_area->width - calc_width);
+      *x_offset = (int)(cell->xalign * (cell_area->width - calc_width));
       *x_offset = MAX(*x_offset, 0);
     }
 
     if (y_offset)
     {
-      *y_offset = cell->yalign * (cell_area->height - calc_height);
+      *y_offset = (int)(cell->yalign * (cell_area->height - calc_height));
       *y_offset = MAX(*y_offset, 0);
     }
   }
